+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed a bug preventing the conference updates from the query builder window.
 - Conference duplications now omit the closure field as intended.
 - Fixed a bug causing the program to crash when exporting data from the query builder if there was an invalid Excel tab name.
+- Fixed a bug where the data pane's update context menu button was disabled if the user lacked delete permissions.
   
 ## 1.0.2
 ### Changed